Move obtain_current_passphrase next to KeyringWrapper.
